Fixing incorrect percent calculation
Fixing incorrect percent calculation due to incorrect casting of data.

+++ src/main/java/net/vortexdata/netwatchdog/modules/northstar/NorthstarRegister.java
@@ -43,7 +43,7 @@ public class NorthstarRegister {
                 successful++;
         }
 
-        return (int) (successful / northstars.size()) * 100;
+        return (int) ((successful / (double) northstars.size()) * 100);
     }
 
     public ArrayList<NorthstarBase> getNorthstars() {
